added base module

+++ include/poolablecomparision.h
@@ -11,7 +11,7 @@ public:
 	bool operator()(Poolable* p1, Poolable* p2) const;
 
 private:
-	bool LessThan(TaskPriority tp1, TaskPriority tp2) const { return (tp1<tp2); }
+	bool LessThan(TaskPriority tp1, TaskPriority tp2) const { return ((int)tp1<(int)tp2); }
 };
 
 #endif // poolablecomparision_h__
